requirements.txt
fastapi
uvicorn
moviepy
requests
python-dotenv
torchaudio
transformers
openai-whisper
langchain
langchain-google-genai
pydantic
aiofiles
aiohttp